% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grab_args.R
\name{grab_args}
\alias{grab_args}
\title{Arg helper}
\usage{
grab_args()
}
\value{
The args passed into a function
}
\description{
\code{grab_args}
}
\details{
This function is intended as a helper to be used within functions. It creates
   an object which stores all the arguments passed into the function/parent
   environment. Useful for when a function has many arguments or \code{...}
   parameter in which named referencing can be too much work.
}
\examples{
.examp <- function(arg = NULL, a = 3, is_fun = TRUE, ...){
  fargs <- grab_args()
  return(fargs)
}

> .examp(test_arg = 123)
$arg
NULL
$a
[1] 3
$is_fun
[1] TRUE
$test_arg
[1] 123

}
\author{
Carl Boneri, \email{carl.boneri@whyles.com}
}
