% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grab_args.R
\name{grab_args_clean}
\alias{grab_args_clean}
\title{Arg helper with cleaning}
\usage{
grab_args_clean()
}
\description{
\code{grab_args_clean}
}
\details{
This is a version of \code{grab_args} that will drop NULL values, as well
  as match any arguments which may have multiple options. The main difference
  here is that this will handle any need for \code{match.arg} in parameter inputs.
}
\examples{
t.search_test2 <- function(q = NULL, geocode = NULL, lang = NULL, locale = NULL, result_type = c("mixed", "recent", "popular"),
                           count = 100, until = NULL, since_id = NULL, max_id = NULL, include_entities = FALSE){

   query <- grab_args_clean()


   query

}

> t.search_test2()
$result_type
[1] "mixed"
$count
[1] 100
$include_entities
[1] FALSE


}
